backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
sqlalchemy>=2.0.29
asyncpg>=0.29.0
pgvector>=0.2.5
pydantic>=2.7.0
pydantic-settings>=2.2.1
redis>=5.0.3
celery>=5.3.6
python-multipart>=0.0.9
email-validator>=2.1.1
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
google-generativeai>=0.5.0
beautifulsoup4>=4.12.3
requests>=2.31.0
pdfplumber>=0.11.0
python-docx>=1.1.0
tenacity>=8.2.3
sentry-sdk>=1.45.0
alembic
